requirements.txt
setuptools>=65.5.0
numpy==1.26.4
torch==2.6.0
torchaudio==2.6.0
transformers==4.46.3
diffusers==0.29.0
librosa==0.11.0
s3tokenizer
resemble-perth==1.0.1
conformer==0.3.2
safetensors==0.5.3
soundfile==0.12.1
tensorboard==2.15.1
wandb==0.16.1
datasets==2.16.0
accelerate==0.25.0
pydub==0.25.1
tqdm==4.66.1
jiwer==3.0.3
phonemizer==3.2.1
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pyyaml==6.0.1

